Add rename profileSlice to myFeedbacksSlice
