+++ client/src/app/store/index.ts
@@ -2,13 +2,13 @@ import { configureStore } from "@reduxjs/toolkit";
 import { useDispatch } from "react-redux";
 import authSlice from "./slice/auth/authSlice";
 import categoriesSlice from "./slice/categories/categoriesSlice";
-import profileSlice from "./slice/profile/profileSlice";
+import myFeedbacksSlice from "./slice/myFeedbacks/myFeedbacksSlice";
 
 export const store = configureStore({
   reducer: {
     auth: authSlice,
     categories: categoriesSlice,
-    profile: profileSlice,
+    myFeedbacks: myFeedbacksSlice,
   },
 });
